+++ src/engine/inmemory-fsprovider.ts
@@ -1,0 +1,36 @@
+
+import * as vscode from "vscode";
+export class MemFS implements vscode.FileSystemProvider {
+	private _emitter = new vscode.EventEmitter<vscode.FileChangeEvent[]>();
+	readonly onDidChangeFile: vscode.Event<vscode.FileChangeEvent[]> = this._emitter.event;
+
+  watch(uri: vscode.Uri, options: { readonly recursive: boolean; readonly excludes: readonly string[]; }): vscode.Disposable {
+		return new vscode.Disposable(() => { });
+  }
+  stat(uri: vscode.Uri): vscode.FileStat | Thenable<vscode.FileStat> {
+    return vscode.workspace.fs.stat(vscode.Uri.from({scheme: uri.fragment, path: uri.path}));
+  }
+  readDirectory(uri: vscode.Uri): [string, vscode.FileType][] | Thenable<[string, vscode.FileType][]> {
+    throw new Error("Method not implemented.");
+  }
+  createDirectory(uri: vscode.Uri): void | Thenable<void> {
+    throw new Error("Method not implemented.");
+  }
+  readFile(uri: vscode.Uri): Uint8Array | Thenable<Uint8Array> {
+    // if( uri )	{
+    //   throw vscode.FileSystemError.FileNotFound();
+    // }
+    return vscode.workspace.fs.readFile(vscode.Uri.from({scheme: uri.fragment, path: uri.path}));
+  }
+  writeFile(uri: vscode.Uri, content: Uint8Array, options: { readonly create: boolean; readonly overwrite: boolean; }): void | Thenable<void> {
+  }
+  delete(uri: vscode.Uri, options: { readonly recursive: boolean; }): void | Thenable<void> {
+    throw new Error("Method not implemented.");
+  }
+  rename(oldUri: vscode.Uri, newUri: vscode.Uri, options: { readonly overwrite: boolean; }): void | Thenable<void> {
+    throw new Error("Method not implemented.");
+  }
+  copy?(source: vscode.Uri, destination: vscode.Uri, options: { readonly overwrite: boolean; }): void | Thenable<void> {
+    throw new Error("Method not implemented.");
+  }
+}