+++ src/engine/search-engine.ts
@@ -1,32 +1,31 @@
-import { IPHtmlElement, IPHtmlNode } from "html-parser/dist/interface";
-import * as ts from "typescript";
 import * as vm from "vm";
 import * as vscode from "vscode";
 import { SerachResult, SerachResultItem } from "../model/search-result.model";
 import {
-  ReplaceDocument,
   ReplaceEdit
 } from "./replace-edit";
+import { ExecuteModes } from "../constants";
+import { QSNode, SearchContext } from "../model/search-context.model";
 
-export type QSNode = any | IPHtmlNode | IPHtmlElement | ts.Node;
 
-export abstract class SearchEngine<TNode> {
-  abstract canApply(uri : vscode.Uri): boolean;
-  abstract validateSearchContext(queryExpr: SearchContext): boolean;
+export abstract class SearchEngine {
+  abstract canApply(uri: vscode.Uri): boolean;
+  abstract validateSearchContext(searchContext: SearchContext): boolean;
   protected abstract searchHtml(
     content: string,
-    queryExpr: SearchContext
-  ): TNode[];
-  abstract getReplacedText(node: TNode): string;
+    searchContext: SearchContext
+  ): QSNode[];
+  abstract getReplacedText(node: QSNode): string;
 
   search(
-    content$: ReplaceDocument,
-    queryExpr: SearchContext
+    content$: vscode.TextDocument,
+    searchContext: SearchContext
   ): SerachResult | null {
     const content = content$.getText();
-    const result = this.searchHtml(content, queryExpr);
+    const result = this.searchHtml(content, searchContext);
     if (result?.length > 0) {
-      const r = new SerachResult(content$.uri, result);
+      const executeMode = searchContext.replaceToggle ? ExecuteModes.replace : ExecuteModes.search;
+      const r = new SerachResult(content$, result, searchContext, executeMode);
       return r;
     }
     return null;
@@ -37,50 +36,39 @@ export abstract class SearchEngine<TNode> {
     replaceExpr: string,
     edit: ReplaceEdit
   ) {
-    try {
-      const uri = searchResult.resourceUri!;
-      const document = await edit.openTextDocument(uri!);
-      for (const item of searchResult.items) {
-        await this._replaceItem(item, document, replaceExpr, edit);
-      }
-
-      await edit.applyEdit();
-    } catch (e: any) {
-      await vscode.window.showErrorMessage(
-        "failed to replace: \n" + e?.toString()
-      );
-      console.log(e, e.stack);
+    const uri = searchResult.resourceUri!;
+    const document = await edit.openTextDocument(uri!);
+    if (document.version !== searchResult.version ){
+      return;
     }
+    for (const item of searchResult.items) {
+      await this._replaceItem(item, document, replaceExpr, edit);
+    }
+
+    await edit.applyEdit();
   }
 
   private async _replaceItem(
     item: SerachResultItem,
-    document: ReplaceDocument,
+    document: vscode.TextDocument,
     replaceExpr: string,
     edit: ReplaceEdit
   ) {
     const baseTag = item.tag;
 
     const $ = this.createClone(baseTag);
+    const original = this.getReplacedText($);
     const vmContext = { $: $ };
     vm.createContext(vmContext);
     const result = vm.runInContext(replaceExpr, vmContext, {
       timeout: 1000,
     });
-
-    await edit.replace(document.uri, this.getRange(baseTag), this.getReplacedText($));
+    const modified = this.getReplacedText($);
+    if (original !== modified) {
+      await edit.replace(document.uri, this.getRange(baseTag), modified);
+    }
   }
-  abstract createClone(v: TNode) :TNode;
+  abstract createClone(v: QSNode): QSNode;
 
-  abstract getRange(baseTag: TNode): readonly [number, number] ;
+  abstract getRange(baseTag: QSNode): readonly [number, number];
 }
-
-export type SearchContext = {
-  search: string;
-  replace?: string;
-  replaceToggle?: boolean;
-  filterToggle?: boolean;
-  includes?: string;
-  excludes?: string;
-};
-