+++ src/test/suite/replace-script-all.test.ts
@@ -4,7 +4,6 @@ import { after } from "mocha";
 import * as vscode from "vscode";
 import { SearchResultPanelProvider } from "../../view/search-result-panel";
 import { NodeHtmlParserAdaptor } from "../../engine/node-html-parser";
-import { ReplaceDocument } from "../../engine/replace-edit";
 // import * as myExtension from '../extension';
 
 suite("ReplaceAll Script Test", () => {
@@ -21,7 +20,7 @@ suite("ReplaceAll Script Test", () => {
   });
 
   async function assertReplace(
-    document: ReplaceDocument,
+    document: vscode.TextDocument,
     search: string,
     replace: string,
     expected: unknown
@@ -29,7 +28,7 @@ suite("ReplaceAll Script Test", () => {
 		const searchContext = {search, replace};
 		const result = new NodeHtmlParserAdaptor().search(document, searchContext );
 
-    await testee.replaceAll(result!, replace);
+    await testee.replaceAll(result!);
     assert.equal(document.getText(), expected);
   }
 
@@ -39,7 +38,7 @@ suite("ReplaceAll Script Test", () => {
         "<ul><li>file</li><li>file</li></ul>" +
         "<ul><li>file</li><li>file</li></ul>",
     });
-    const queryExpr= "ul:has(li)";
+    const searchContext= "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -49,7 +48,7 @@ suite("ReplaceAll Script Test", () => {
       "<ul><li>file</li></ul><li>file</li>"+ 
 			"<ul><li>file</li></ul><li>file</li>";
 
-			await assertReplace(document, queryExpr, replaceExpr, expected);
+			await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("remove and insert to AfterEnd, multiline", async () => {
@@ -65,7 +64,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -82,7 +81,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul><li>file</li>
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   // XXX closing/empty tag
@@ -95,7 +94,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr= "ul:has(li)";
+    const searchContext= "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -108,7 +107,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul><li />
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test.skip("remove and insert to AfterEnd, preserve closing/empty tag", async () => {
@@ -120,7 +119,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr= "ul:has(li)";
+    const searchContext= "ul:has(li)";
     const replaceExpr = `
 		`;
     const expected = `
@@ -130,7 +129,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
 
@@ -145,7 +144,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr= "ul:has(li)";
+    const searchContext= "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -160,7 +159,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul><li b="1" a="2"></li>
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("setAttribute", async () => {
@@ -174,7 +173,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr= "ul:has(li)";
+    const searchContext= "ul:has(li)";
     const replaceExpr = `
 			$.setAttribute("addAttr", "new");
 		`;
@@ -187,7 +186,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
 	test("nested", async () => {
@@ -208,7 +207,7 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`,
 		});
-		const queryExpr = "ul:has(li)";
+		const searchContext = "ul:has(li)";
 		const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -228,6 +227,6 @@ suite("ReplaceAll Script Test", () => {
 		</ul>
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
 	});
 });