+++ src/test/suite/replace-script.test.ts
@@ -3,7 +3,6 @@ import { after } from "mocha";
 
 import * as vscode from "vscode";
 import { NodeHtmlParserAdaptor } from "../../engine/node-html-parser";
-import { ReplaceDocument } from "../../engine/replace-edit";
 import { SearchResultPanelProvider } from "../../view/search-result-panel";
 
 suite("Replace Script Test", () => {
@@ -20,7 +19,7 @@ suite("Replace Script Test", () => {
   });
 
   async function assertReplace(
-    document: ReplaceDocument,
+    document: vscode.TextDocument,
     search: string,
     replace: string,
     expected: unknown
@@ -28,7 +27,7 @@ suite("Replace Script Test", () => {
 		const searchContext = {search, replace};
 		const result = new NodeHtmlParserAdaptor().search(document, searchContext );
 
-		await testee.replace(result!.items[0], replace);
+		await testee.replace(result!.items[0]);
     assert.equal(document.getText(), expected);
   }
 
@@ -36,19 +35,19 @@ suite("Replace Script Test", () => {
     const document = await vscode.workspace.openTextDocument({
       content: "<ul><li>file</li></ul>",
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 		`;
     const expected = "<ul><li>file</li></ul>";
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("remove and insert to AfterEnd", async () => {
     const document = await vscode.workspace.openTextDocument({
       content: "<ul><li>file</li></ul>",
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -56,7 +55,7 @@ suite("Replace Script Test", () => {
 		`;
     const expected = "<ul></ul><li>file</li>";
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("remove and insert to AfterEnd, multiline", async () => {
@@ -67,7 +66,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -79,7 +78,7 @@ suite("Replace Script Test", () => {
 		</ul><li>file</li>
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   // TODO self closing/empty tag become decomposite by parser.
@@ -92,7 +91,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -105,7 +104,7 @@ suite("Replace Script Test", () => {
 		</ul><li />
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   // TODO self closing/empty tag become decomposite by parser.
@@ -117,7 +116,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -129,7 +128,7 @@ suite("Replace Script Test", () => {
 		</ul><li b="1" a="2"></li>
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("modify attributes", async () => {
@@ -140,7 +139,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 			$.setAttribute("appendttr", "APPENDED");
 			$.setAttribute("modifyAttr", "NEW_VALUE");
@@ -152,7 +151,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("preserve comment block", async () => {
@@ -167,7 +166,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 		var s = $.querySelector("li");
 		$.removeChild(s);
@@ -183,7 +182,7 @@ suite("Replace Script Test", () => {
 		</ul><li><!-- second --></li>
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("preserve multiline attribute", async () => {
@@ -197,7 +196,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)";
+    const searchContext = "ul:has(li)";
     const replaceExpr = `
 		var s = $.querySelector("li");
 		$.removeChild(s);
@@ -212,7 +211,7 @@ suite("Replace Script Test", () => {
 		'></li>
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
   test("preserve multiline attribute", async () => {
@@ -226,7 +225,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
     });
-    const queryExpr = "ul:has(li)" ;
+    const searchContext = "ul:has(li)" ;
     const replaceExpr = `
 		$.setAttribute("replaced", "")
 		`;
@@ -239,7 +238,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`;
 
-    await assertReplace(document, queryExpr, replaceExpr, expected);
+    await assertReplace(document, searchContext, replaceExpr, expected);
   });
 
 	test("nested", async () => {
@@ -260,7 +259,7 @@ suite("Replace Script Test", () => {
 		</ul>
 		`,
 		});
-		const queryExpr = "ul:has(li)";
+		const searchContext = "ul:has(li)";
 		const replaceExpr = `
 			var s = $.querySelector("li");
 			$.removeChild(s);
@@ -280,6 +279,6 @@ suite("Replace Script Test", () => {
 		</ul>
 		`;
 
-		await assertReplace(document, queryExpr, replaceExpr, expected);
+		await assertReplace(document, searchContext, replaceExpr, expected);
 	});
 });